Upgrade to .NET 8.0
