+++ .gitignore
@@ -395,6 +395,7 @@ FodyWeavers.xsd
 *.msp
 
 # JetBrains Rider
+.idea/
 *.sln.iml
 
 /BenchmarkDotNet.Artifacts
