+++ src/MiniExcel/MiniExcelLibs.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net45;netstandard2.0;net6.0;</TargetFrameworks>
+		<TargetFrameworks>net45;netstandard2.0;net8.0;</TargetFrameworks>
 		<Version>1.34.2</Version>
 	</PropertyGroup>
 	<PropertyGroup>
