+++ tests/MiniExcelTests/MiniExcelTests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net6.0</TargetFrameworks>
+		<TargetFrameworks>net8.0</TargetFrameworks>
 
 		<IsPackable>false</IsPackable>
 		<AssemblyOriginatorKeyFile>miniexcel.snk</AssemblyOriginatorKeyFile>
